requirements.txt
absl-py==2.3.0
altair==5.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs @ file:///C:/b/abs_89hmquz5ga/croot/attrs_1734533130810/work
beautifulsoup4 @ file:///C:/b/abs_d5wytg_p0w/croot/beautifulsoup4-split_1718029833749/work
blinker==1.9.0
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
certifi==2025.4.26
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
comm==0.2.2
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.8.6
decorator==5.1.1
executing==2.1.0
fastjsonschema @ file:///C:/b/abs_4ev90296ly/croot/python-fastjsonschema_1731939386416/work
flatbuffers==25.2.10
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h11 @ file:///C:/b/abs_1czwoyexjf/croot/h11_1706652332846/work
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
ipywidgets @ file:///C:/b/abs_f03en2fv37/croot/ipywidgets_1733504604628/work
jedi==0.19.1
Jinja2==3.1.6
joblib @ file:///C:/b/abs_f4b98l6lgk/croot/joblib_1718217224240/work
jsonschema @ file:///C:/b/abs_394_t6__xq/croot/jsonschema_1728486718320/work
jsonschema-specifications @ file:///C:/Users/dev-admin/py312/jsonschema-specifications_1706803038066/work
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets @ file:///C:/b/abs_70q5kcqvoa/croot/jupyterlab_widgets_1733441048419/work
keras==3.10.0
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.38.2
nbformat @ file:///C:/b/abs_c2jkw46etm/croot/nbformat_1728050303821/work
nest-asyncio==1.6.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.16.0
outcome @ file:///C:/b/abs_774iyc6vgh/croot/outcome_1741014061865/work
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
parso==0.8.4
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs==4.3.6
plotly @ file:///C:/b/abs_28qon8a90a/croot/plotly_1740385581534/work
prompt_toolkit==3.0.48
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
Pygments @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pygments_1699474141968/work
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt6==6.7.1
PyQt6_sip @ file:///C:/b/abs_28s7k4h_hl/croot/pyqt-split_1740498234166/work/pyqt_sip
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==306
pyzmq==26.2.0
referencing @ file:///C:/Users/dev-admin/py312/referencing_1706802962559/work
requests==2.32.3
rich==14.0.0
rpds-py @ file:///C:/b/abs_0c6z5kcdb6/croot/rpds-py_1736545465023/work
scikit-learn @ file:///C:/b/abs_0dwalqot3o/croot/scikit-learn_1737988780693/work/dist/scikit_learn-1.6.1-cp312-cp312-win_amd64.whl#sha256=182506dc5e8dc16a17946480ff4567ca27142b879c588302c4fb409e5efadb75
scipy @ file:///C:/b/abs_57lp_c7nt4/croot/scipy_1743153260151/work/dist/scipy-1.15.2-cp312-cp312-win_amd64.whl#sha256=48c87ee9794dea0f7465e0cbfce2531f658de9035b858d986e84680cc495b1ea
seaborn @ file:///C:/b/abs_522_0_dy9p/croot/seaborn_1741187268016/work
selenium @ file:///C:/b/abs_0fryk0y0zp/croot/selenium_1727169285181/work
setuptools==80.9.0
sip @ file:///C:/b/abs_5cto136kse/croot/sip_1738856220313/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
sortedcontainers @ file:///tmp/build/80754af9/sortedcontainers_1623949099177/work
soupsieve @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/soupsieve_1699496611169/work
stack-data==0.6.3
streamlit==1.45.0
tenacity @ file:///C:/b/abs_1749lsqys6/croot/tenacity_1730304400080/work
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.18.1
termcolor==3.1.0
threadpoolctl @ file:///C:/b/abs_def0dwqlft/croot/threadpoolctl_1719407816649/work
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
trio @ file:///C:/b/abs_828y0nwq4i/croot/trio_1729106637704/work
trio-websocket @ file:///C:/b/abs_45nc1akikg/croot/trio-websocket_1741025743389/work
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urljoin==1.0.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client @ file:///C:/b/abs_5dmnxxoci9/croot/websocket-client_1715878351319/work
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension @ file:///C:/b/abs_df0o5rkt3b/croot/widgetsnbextension_1733439593530/work
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt==1.17.2
wsproto @ file:///C:/b/abs_31cskpy2sd/croot/wsproto_1725051559217/work
